fix weatherforecast, aggiunta microservizio notifiche con gestione SQS, aggiunta parte conferma invito su visitmanager con get di visite/inviti

+++ frontend/model/WeatherForecast.go
@@ -1,19 +1,19 @@
 package model
 
 type WeatherForecasts struct {
-	Date        []string `json:"time"`
-	Temperature []int    `json:"temperature"`
-	Cloudy      []int    `json:"cloud_cover"`
-	Rainy       []int    `json:"precipitation"`
-	Wind        []int    `json:"wind_speed"`
-	Humidity    []int    `json:"humidity"`
+	Date        []string  `json:"time"`
+	Temperature []float32 `json:"temperature"`
+	Cloudy      []int     `json:"cloud_cover"`
+	Rainy       []int     `json:"precipitation"`
+	Wind        []float32 `json:"wind_speed"`
+	Humidity    []int     `json:"humidity"`
 }
 
 type WeatherForecast struct {
 	Date        string
-	Temperature int
+	Temperature float32
 	Cloudy      int
 	Rainy       int
-	Wind        int
+	Wind        float32
 	Humidity    int
 }
